I've added a few small tests just to assert the connectedness of the database and make sure the data collection methods are all working correctly. I'm starting to get to grips more with JDBC but to be honest it seems a little clunky on the whole, missing a lot of seemingly obvious methods.
Now that I'm more comfortable with the general SQL/Java interface I'm
going to construct the object which will hold all the necessary client
data to fill the form. Owing to my earlier discovery of the inherent
problems of moving a result set around I think I'm going to create it in
the entry package, populate it in the Database package and finally pass
it throught to the form filling package where I am considering the use a
chain of responisbility for filling each page of the form in turn.